(more of andre's cleanups)
More or less a trivial cleanup that reduces lines of code quite a bit
while improving readability.

The only nontrivial change is the skip of the register keyword for
the diff variable. Nowadays, register should not be used anymore as
it does more harm than good. We're using -O anyway.