Disable director in all swig files. This has big impact on
a) speed
b) wrapper code size
c) inheritance.

Good news first, every virtual function should be *much faster* now, as the
call is not anymore rerouted via the director wrapper to python and then from
there to the correct c++ function. Furthermore the wrapper code size decreases
a lot, the inheritance chain is shorter and compile times are reduced.

The sad news is that virtual functions that *were overloaded via python code*
will not be used from inside the C++ code (not sure if anyone expected this to
		work - so it might not actually be that bad).