This attempts to get rid of the language specific code in lib/io.{h,cpp} and lib/Signal.{cpp,h}.
To this end every interface has to define a couple of global functions called
sg_print_{message,warning,error} and sg_cancel_computations. The printing
functions are called whenever anything is writting to the screen (and don't
raise exceptions or anything).

The cancel function is called to allow an interface to check for the occurrance
of a signal and to set the immediate cancel flag. It will be called from
core/busy loops only. This is to fix the varius segmentation faults caused by
throwing exceptions within signal handlers (which are illegal - not in the
standard).