use %import to make other data types known to swig
To this end we need to introduce *_includes.i for the respective
includes that can then be re-used and %included.

For example for Features we put includes in Features_includes.i
and issue in Features.i a %include "Features_includes.i".

The Kernel module needs features and thus make them known. So we need to

%include "Features_includes.i" /* to include the C/C++ headers */
%import "Features.i" /* to make the data types known to swig */