Put implementations of templated functions into .cpp files
This is possible when strictly following
http://www.parashift.com/c++-faq-lite/templates.html#faq-35.15

i.e. we must know beforehand which types we want to support and specify
them like

template class Foo<int32_t>;

at the end of the .cpp file. This all holds true for feature classes
and should enable us to move all but definitions from .h files to
.cpp's.