add custom reduce_ex and reconstructor magic functions for (un)pickling
Whenever an object is of type SGObject we store 'modshogun.Type' as
type instead of the class type. The reconstructor is patched to cope
with that. This will change the pickle format of any python program
shogun writes - however this is a compatible change when going shogun
3.1 -> 3.2 but not downwards compatible.

This works around a python3.3.X pickle issue where shogun types are
wrongly searched for in imp.Type instead of imp.modshogun as used to
work with python 2 up to python 3.2.