Cleanup broken links and fix some issues on Shogun's webpages. (#4516)
* Remove the link to the old team page.

Fix #4512.

* Fix deadlinks in INTERFACES.md

For #4512

* Use inline script tag for custom LaTeX commands. (#4512)

Previoulsy, we load MathJax and mathconfig.js asynchronously. This is
not good for mathconfig.js because it requires the main variable for
MathJax.js.

* Fix style of reference section. (#4512)

Note: the problem seems to originally come from boostrap and how
our biblatex plugin uses `label class` to generates reference entries.

* Update links and some installation information. (#4512)

The information includes Docker and MacOS brew installation.

* Remove duplicated years in references.bib. (#4512)

"Gaussian Processes for Machine Learning" has duplicate years and causes
an error when generating the cookbook webpages. According to the book's
webpage, the published year is 2006.

* Remove duplicate import statements for shogun-style.css (#4512).

* Revert "Use inline script tag for custom LaTeX commands. (#4512)"

This reverts commit 4413403e2ad014f349c1f2de98ca615d828f0f7e.

* set mathjax configuration using window.MathJax (#4512)

* use relative urls for links to documents. (#4512)